Dialog show two images: file sizes added

+++ pordb.py
@@ -614,6 +614,7 @@ class MeinDialog(QtGui.QMainWindow, MainWindow):
 		self.ausgabe_in_table()
 		self.bilder_aktuell()
 		self.suchfeld.setFocus()
+	# end of tableWidgetBilderdropEvent
 				
 	def onPageFirst(self):
 		self.start_bilder = 0
@@ -3073,23 +3074,25 @@ class MeinDialog(QtGui.QMainWindow, MainWindow):
 				if os.path.exists(datei_alt) and os.path.exists(datei_neu) and datei_alt != datei_neu:
 					dialog = ShowTwoImages(datei_alt, datei_neu)
 					app.restoreOverrideCursor()
-					dialog.exec_()
-					datei = dialog.datei()
-					if datei == 1:
-						try:
-							os.remove(datei_neu)
-							os.rename(datei_alt, datei_neu)
-						except:
-							message = QtGui.QMessageBox.critical(self, self.trUtf8("Error "), self.trUtf8("Image file could not be renamed"))
-					elif datei == 2:
-						try:
-							os.remove(datei_alt)
-						except:
-							message = QtGui.QMessageBox.critical(self, self.trUtf8("Error "), self.trUtf8("Image file could not be renamed"))
+					if dialog.exec_():
+						datei = dialog.datei()
+						if datei == 1:
+							try:
+								os.remove(datei_neu)
+								os.rename(datei_alt, datei_neu)
+							except:
+								message = QtGui.QMessageBox.critical(self, self.trUtf8("Error "), self.trUtf8("Image file could not be renamed"))
+						elif datei == 2:
+							try:
+								os.remove(datei_alt)
+							except:
+								message = QtGui.QMessageBox.critical(self, self.trUtf8("Error "), self.trUtf8("Image file could not be renamed"))
+						else:
+							message = QtGui.QMessageBox.information(self, self.trUtf8("Information "), self.trUtf8("Renaming canceled"))
+							self.suchfeld.setCurrentIndex(-1)
+							self.suchfeld.setFocus()
+							return
 					else:
-						message = QtGui.QMessageBox.information(self, self.trUtf8("Information "), self.trUtf8("Renaming canceled"))
-						self.suchfeld.setCurrentIndex(-1)
-						self.suchfeld.setFocus()
 						return
 				else:
 					try:
